shared/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge1-shared"
version = "0.1.0"
description = "Shared utilities and testing harness for Forge 1 monorepo"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Forge 1 Team" }
]
dependencies = [
  "pydantic>=2.8,<3",
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "mypy>=1.10",
  "ruff>=0.5",
  "black>=24.4",
  "types-PyYAML",
]

[tool.setuptools.packages.find]
include = ["shared*"]

[tool.setuptools.package-data]
"shared.testing.suites" = ["*.yaml"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "BLE", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = false


